requirements.txt
setuptools>=65.5.1
crewai==0.11.0
langchain-core==0.1.17
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1
deep-translator==1.11.4
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.31.1
chromadb==0.4.22
sentence-transformers==2.5.1
